refactored Template Link table: - filters are separated from "add new" row - added icons indicating the purpose of dropdowns

+++ src/main/java/com/soam/web/templatelink/TemplateLinkController.java
@@ -12,6 +12,7 @@ import com.soam.model.templatelink.TemplateLinkRepository;
 import jakarta.validation.Valid;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
@@ -63,49 +64,51 @@ public class TemplateLinkController {
         return objectiveTemplateRepository.findAllByOrderByName();
     }
 
-    @ModelAttribute("newTemplateLink")
-    public TemplateLink populateNewTemplateLink(TemplateLink newTemplateLink) {
-        return newTemplateLink;
-    }
-
     @ModelAttribute("templateLinks")
-    public Iterable<TemplateLink> populateTemplateLinks(@ModelAttribute("newTemplateLink") TemplateLink newTemplateLink) {
-        if (newTemplateLink.getSpecificationTemplate() != null && newTemplateLink.getStakeholderTemplate() != null) {
+    public Iterable<TemplateLink> populateTemplateLinks(@ModelAttribute("templateLinkForm") TemplateLinkFormDto templateLinkForm) {
+        if (templateLinkForm.getFilterSpecificationTemplate() != null && templateLinkForm.getFilterStakeholderTemplate() != null) {
             return templateLinkRepository.findBySpecificationTemplateAndStakeholderTemplate(
-                    newTemplateLink.getSpecificationTemplate(), newTemplateLink.getStakeholderTemplate(), TEMPLATE_LINK_SORT);
-        } else if (newTemplateLink.getSpecificationTemplate() != null && newTemplateLink.getStakeholderTemplate() == null) {
-            return templateLinkRepository.findBySpecificationTemplate(newTemplateLink.getSpecificationTemplate(), TEMPLATE_LINK_SORT);
-        } else if (newTemplateLink.getSpecificationTemplate() == null && newTemplateLink.getStakeholderTemplate() != null) {
-            return templateLinkRepository.findByStakeholderTemplate(newTemplateLink.getStakeholderTemplate(), TEMPLATE_LINK_SORT);
+                    templateLinkForm.getFilterSpecificationTemplate(), templateLinkForm.getFilterStakeholderTemplate(), TEMPLATE_LINK_SORT);
+        } else if (templateLinkForm.getFilterSpecificationTemplate() != null && templateLinkForm.getFilterStakeholderTemplate() == null) {
+            return templateLinkRepository.findBySpecificationTemplate(templateLinkForm.getFilterSpecificationTemplate(), TEMPLATE_LINK_SORT);
+        } else if (templateLinkForm.getFilterSpecificationTemplate() == null && templateLinkForm.getFilterStakeholderTemplate() != null) {
+            return templateLinkRepository.findByStakeholderTemplate(templateLinkForm.getFilterStakeholderTemplate(), TEMPLATE_LINK_SORT);
         } else {
             return templateLinkRepository.findAll(TEMPLATE_LINK_SORT);
         }
     }
 
-    @ModelAttribute("isFiltered")
-    public boolean populatedIsFiltered(TemplateLink newTemplateLink) {
-        return newTemplateLink.getSpecificationTemplate() != null || newTemplateLink.getStakeholderTemplate() != null;
-    }
-
     @RequestMapping(value = "/templateLink/list", method = {RequestMethod.GET, RequestMethod.POST})
-    public String listTemplateLinks() {
+    public String listTemplateLinks(@ModelAttribute("templateLinkForm") TemplateLinkFormDto templateLinkForm, Model model) {
+        if (templateLinkForm.getNewTemplateLink() != null) {
+            if (templateLinkForm.getFilterSpecificationTemplate() != null) {
+                templateLinkForm.getNewTemplateLink().setSpecificationTemplate(
+                        templateLinkForm.getFilterSpecificationTemplate());
+            }
+            if (templateLinkForm.getFilterStakeholderTemplate() != null) {
+                templateLinkForm.getNewTemplateLink().setStakeholderTemplate(
+                        templateLinkForm.getFilterStakeholderTemplate());
+            }
+        }
         return VIEWS_TEMPLATE_LINK_LIST;
     }
 
     @PostMapping("/templateLink/new")
-    public String processCreateForm(@Valid TemplateLink templateLink, BindingResult bindingResult,
+    public String processCreateForm(@Valid TemplateLinkFormDto templateLinkForm, BindingResult bindingResult,
                                     RedirectAttributes redirectAttributes) {
-        Optional<TemplateLink> maybeExistingTemplateLink = 
+        TemplateLink templateLink = templateLinkForm.getNewTemplateLink();
+        Optional<TemplateLink> maybeExistingTemplateLink =
                 templateLinkRepository.findBySpecificationTemplateAndStakeholderTemplateAndObjectiveTemplate(
                         templateLink.getSpecificationTemplate(), templateLink.getStakeholderTemplate(),
                         templateLink.getObjectiveTemplate());
-        redirectAttributes.addFlashAttribute("newTemplateLink", templateLink);
+        redirectAttributes.addFlashAttribute("templateLinkForm", templateLinkForm);
         if (maybeExistingTemplateLink.isEmpty()) {
             if (bindingResult.hasErrors()) {
-                redirectAttributes.addFlashAttribute(Util.DANGER, "Specify all fields for the new template link.");
+                //the UI should never cause this error. This is protection mostly from malformed programmatic POST
+                redirectAttributes.addFlashAttribute(Util.DANGER, "New template link data is not complete.");
             } else {
                 templateLinkRepository.save(templateLink);
-                redirectAttributes.addFlashAttribute(Util.SUCCESS, String.format("Successfully created %s.", getTemplateLinkTitle(templateLink)));
+                redirectAttributes.addFlashAttribute(Util.SUB_FLASH, String.format("Successfully created template link %s.", getTemplateLinkTitle(templateLink)));
             }
         } else {
             redirectAttributes.addFlashAttribute(Util.DANGER, String.format("Template link %s already exists.", getTemplateLinkTitle(templateLink)));
@@ -113,16 +116,22 @@ public class TemplateLinkController {
         return REDIRECT_TEMPLATE_LINK_LIST;
     }
 
-    @PostMapping("/templateLink/{templateLinkId}/delete")
-    public String processDeleteTemplateLink(@PathVariable("templateLinkId") int templateLinkId,
+    @PostMapping("/templateLink/delete")
+    public String processDeleteTemplateLink(@Valid TemplateLinkFormDto templateLinkForm, BindingResult bindingResult,
                                             RedirectAttributes redirectAttributes) {
-        Optional<TemplateLink> maybeTemplateLink = templateLinkRepository.findById(templateLinkId);
+        Optional<TemplateLink> maybeTemplateLink = templateLinkRepository.findById(templateLinkForm.getDeleteTemplateLinkId());
 
+        redirectAttributes.addFlashAttribute("templateLinkForm", templateLinkForm);
         if (maybeTemplateLink.isPresent()) {
-            redirectAttributes.addFlashAttribute(Util.SUB_FLASH, String.format("Successfully deleted %s", getTemplateLinkTitle(maybeTemplateLink.get())));
-            templateLinkRepository.delete(maybeTemplateLink.get());
+            if (bindingResult.hasErrors()) {
+                //the UI should never cause this error. This is protection mostly from malformed programmatic POST
+                redirectAttributes.addFlashAttribute(Util.DANGER, "New form data is malformed.");
+            } else {
+                redirectAttributes.addFlashAttribute(Util.SUB_FLASH, String.format("Successfully deleted template link %s.", getTemplateLinkTitle(maybeTemplateLink.get())));
+                templateLinkRepository.delete(maybeTemplateLink.get());
+            }
         }else{
-            redirectAttributes.addFlashAttribute(Util.DANGER, "Error deleting template link");
+            redirectAttributes.addFlashAttribute(Util.DANGER, "Error deleting template link.");
         }
         return REDIRECT_TEMPLATE_LINK_LIST;
     }
